+++ sdk/clientcore/core/src/main/java/io/clientcore/core/implementation/instrumentation/otel/tracing/OTelContext.java
@@ -6,9 +6,9 @@ package io.clientcore.core.implementation.instrumentation.otel.tracing;
 import io.clientcore.core.implementation.ReflectiveInvoker;
 import io.clientcore.core.implementation.instrumentation.otel.FallbackInvoker;
 import io.clientcore.core.implementation.instrumentation.otel.OTelInitializer;
+import io.clientcore.core.instrumentation.InstrumentationContext;
 import io.clientcore.core.instrumentation.tracing.TracingScope;
-import io.clientcore.core.instrumentation.tracing.SpanKind;
-import io.clientcore.core.util.ClientLogger;
+import io.clientcore.core.instrumentation.logging.ClientLogger;
 
 import static io.clientcore.core.implementation.ReflectionUtils.getMethodInvoker;
 import static io.clientcore.core.implementation.instrumentation.otel.OTelInitializer.CONTEXT_CLASS;
@@ -24,19 +24,19 @@ class OTelContext {
     private static final FallbackInvoker GET_INVOKER;
 
     // this context key will indicate if the span is created by client core
-    // AND has client or internal kind (logical client operation)
     // this is used to suppress multiple spans created for the same logical operation
     // such as convenience API on top of protocol methods when both as instrumented.
     // We might need to suppress logical server (consumer) spans in the future, but that
     // was not necessary so far
-    private static final Object HAS_CLIENT_SPAN_CONTEXT_KEY;
+    private static final Object CLIENT_CORE_SPAN_CONTEXT_KEY;
 
     static {
         ReflectiveInvoker currentInvoker = null;
         ReflectiveInvoker makeCurrentInvoker = null;
         ReflectiveInvoker withInvoker = null;
         ReflectiveInvoker getInvoker = null;
-        Object hasClientSpanContextKey = null;
+
+        Object clientCoreSpanContextKey = null;
         Object rootContext = null;
 
         if (OTelInitializer.isInitialized()) {
@@ -50,10 +50,11 @@ class OTelContext {
                 ReflectiveInvoker contextKeyNamedInvoker
                     = getMethodInvoker(CONTEXT_KEY_CLASS, CONTEXT_KEY_CLASS.getMethod("named", String.class));
 
-                hasClientSpanContextKey = contextKeyNamedInvoker.invoke("client-core-call");
+                clientCoreSpanContextKey = contextKeyNamedInvoker.invoke("client-core-span");
 
                 ReflectiveInvoker rootInvoker = getMethodInvoker(CONTEXT_CLASS, CONTEXT_CLASS.getMethod("root"));
                 rootContext = rootInvoker.invoke();
+
             } catch (Throwable t) {
                 OTelInitializer.initError(LOGGER, t);
             }
@@ -63,7 +64,7 @@ class OTelContext {
         MAKE_CURRENT_INVOKER = new FallbackInvoker(makeCurrentInvoker, NOOP_SCOPE, LOGGER);
         WITH_INVOKER = new FallbackInvoker(withInvoker, LOGGER);
         GET_INVOKER = new FallbackInvoker(getInvoker, LOGGER);
-        HAS_CLIENT_SPAN_CONTEXT_KEY = hasClientSpanContextKey;
+        CLIENT_CORE_SPAN_CONTEXT_KEY = clientCoreSpanContextKey;
     }
 
     static Object getCurrent() {
@@ -79,21 +80,41 @@ class OTelContext {
         return (AutoCloseable) scope;
     }
 
-    static Object markCoreSpan(Object context, SpanKind spanKind) {
+    static Object markCoreSpan(Object context, OTelSpan span) {
         assert CONTEXT_CLASS.isInstance(context);
-        if (spanKind == SpanKind.CLIENT || spanKind == SpanKind.INTERNAL) {
-            Object updatedContext = WITH_INVOKER.invoke(context, HAS_CLIENT_SPAN_CONTEXT_KEY, Boolean.TRUE);
-            if (updatedContext != null) {
-                return updatedContext;
-            }
-        }
-        return context;
+        Object updatedContext = WITH_INVOKER.invoke(context, CLIENT_CORE_SPAN_CONTEXT_KEY, span);
+        return updatedContext == null ? context : updatedContext;
     }
 
-    static boolean hasClientCoreSpan(Object context) {
+    static OTelSpan getClientCoreSpan(Object context) {
         assert CONTEXT_CLASS.isInstance(context);
-        Object flag = GET_INVOKER.invoke(context, HAS_CLIENT_SPAN_CONTEXT_KEY);
-        assert flag == null || flag instanceof Boolean;
-        return Boolean.TRUE.equals(flag);
+        Object clientCoreSpan = GET_INVOKER.invoke(context, CLIENT_CORE_SPAN_CONTEXT_KEY);
+        assert clientCoreSpan == null || clientCoreSpan instanceof OTelSpan;
+        return (OTelSpan) clientCoreSpan;
+    }
+
+    /**
+     * Get the OpenTelemetry context from the given context.
+     *
+     * @param context the context
+     * @return the OpenTelemetry context
+     */
+    static Object fromInstrumentationContext(InstrumentationContext context) {
+        if (context instanceof OTelSpanContext) {
+            Object otelContext = ((OTelSpanContext) context).getOtelContext();
+            if (otelContext != null) {
+                return otelContext;
+            }
+        }
+
+        Object currentContext = CURRENT_INVOKER.invoke();
+        if (context != null) {
+            Object spanContext = OTelSpanContext.toOTelSpanContext(context);
+            Object span = OTelSpan.wrapSpanContext(spanContext);
+
+            return OTelSpan.storeInContext(span, currentContext);
+        }
+
+        return currentContext;
     }
 }